move to c++20
No reason to not use the benefits of already implemented features, plus
guarantees. For example, in C++20 there is finally only one integer
representation -- two's complement. This means that e.g. range of 8-bit
signed integer is officially -128 to 127, instead of -127 to 127 if we
were accounting for two other integer representations.