redo vector as a matrix, point to contain vector
This was first and foremost done to avoid unnecessary copies during
scale(), translate() etc. packing/unpacking. It also removes a lot of
copypaste code, except that multiplication had to be written thrice: for
matrix, vector and point. But still, now I can also get rid of scale(),
translate() etc. and just use multiplication however I want, in a much
cleaner way. Hopefully this is the last patch that practically rebuilds
the world.